Enable warnings for maintainer mode.  Fix warnings.
* configure.ac: Apply GCC specific options for maintainer mode.
* src/assuan-handler.c, src/assuan-logging.c: Fix warnings.
* src/assuan-socket-connect.c, src/assuan-socket.c: Likewise.
* src/assuan-uds.c, src/debug.c, src/system-posix.c: Likewise.
* src/system-w32.c, src/system.c: Likewise.
* tests/socks5.c (main): Use int for the check of the value 65535.

--

The code in configure.ac is from libgpg-error master.

+++ tests/socks5.c
@@ -207,7 +207,7 @@ main (int argc, char **argv)
     }
   else if (opt_byname)
     {
-      unsigned short port;
+      unsigned int port;
       char *cred;
 
       if (user || pass)
